Restart prefab fix
Restart (als je op R drukt) pauzeert het spel, als je dan 'nee' selecteert gaat het spel weer op play.
