+++ Assets/Scenes/LevelMelvin1.unity
@@ -632,7 +632,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 3335870511896439565, guid: c997599ddcdce5846b6bbabc70c1f8ae, type: 3}
       propertyPath: m_RootOrder
-      value: 4
+      value: 3
       objectReference: {fileID: 0}
     - target: {fileID: 3335870511896439565, guid: c997599ddcdce5846b6bbabc70c1f8ae, type: 3}
       propertyPath: m_LocalPosition.x
@@ -2032,71 +2032,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 149954983652433154, guid: 888a74e5760eb3147a8f77a2e6dacb61, type: 3}
   m_PrefabInstance: {fileID: 1513708518}
   m_PrefabAsset: {fileID: 0}
---- !u!1001 &1450786257
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    serializedVersion: 3
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
-      propertyPath: m_RootOrder
-      value: 3
-      objectReference: {fileID: 0}
-    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 5.9888887
-      objectReference: {fileID: 0}
-    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: -3.0831268
-      objectReference: {fileID: 0}
-    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0.389761
-      objectReference: {fileID: 0}
-    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8061037776434441370, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
-      propertyPath: m_Name
-      value: RestartObject
-      objectReference: {fileID: 0}
-    - target: {fileID: 8061037776622878540, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
-      propertyPath: m_Camera
-      value: 
-      objectReference: {fileID: 2059590899}
-    m_RemovedComponents: []
-    m_RemovedGameObjects: []
-    m_AddedGameObjects: []
-    m_AddedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
 --- !u!1 &1454363924 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 4270832553172412253, guid: f1d9c88f8b606ff45bed382351f4f004, type: 3}
@@ -7310,6 +7245,83 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 3dec2a4c6b2f940919c8cca90801d4b8, type: 3}
+--- !u!1001 &1990014853
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_RootOrder
+      value: 4
+      objectReference: {fileID: 0}
+    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 5.9888887
+      objectReference: {fileID: 0}
+    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: -3.0831268
+      objectReference: {fileID: 0}
+    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0.389761
+      objectReference: {fileID: 0}
+    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8061037776434441369, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8061037776434441370, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_Name
+      value: RestartObject
+      objectReference: {fileID: 0}
+    - target: {fileID: 8061037776622878540, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_Camera
+      value: 
+      objectReference: {fileID: 2059590899}
+    - target: {fileID: 8061037776622878540, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_SortingOrder
+      value: 5
+      objectReference: {fileID: 0}
+    - target: {fileID: 8061037777640986916, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_SortingOrder
+      value: 5
+      objectReference: {fileID: 0}
+    - target: {fileID: 8061037777640986918, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: e8a6a840c6a28184b87243eae5258a67, type: 3}
 --- !u!1001 &2003423538
 PrefabInstance:
   m_ObjectHideFlags: 0
