+++ Assets/Scripts/Management/DisablePause.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b7b99c2655696fe478eee4f608378941
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
