--- Assets/Scripts/Others/ConfirmBoxEnable.cs
@@ -1,15 +1,0 @@
-using UnityEngine;
-
-public class ConfirmBoxEnable : MonoBehaviour
-{
-    [SerializeField] private GameObject ConfirmationBox;
-
-    void Update()
-    {
-        if (Input.GetKey(KeyCode.R))
-        {
-            ConfirmationBox.gameObject.SetActive(true);
-            Time.timeScale = 0;
-        }
-    }
-}
